data-pipeline/requirements.txt

selenium
webdriver-manager
firebase-admin
google-cloud-firestore
